gotranslatebot поддержка чата и конструктора

+++ googletranslate.py
@@ -63,44 +63,48 @@ def main():
             payload = bot.get_payload(last_update)
             callback_id = bot.get_callback_id(last_update)
             mid = bot.get_message_id(last_update)
-            if text == '/lang' or text == '@gotranslatebot /lang':
-                buttons = [[{"type": 'callback',
-                             "text": 'Авто|Auto',
-                             "payload": 'auto'},
-                            {"type": 'callback',
-                             "text": 'Русский',
-                             "payload": 'ru'},
-                            {"type": 'callback',
-                             "text": 'English',
-                             "payload": 'en'}]]
-                bot.send_buttons('Направление перевода\nTranslation direction', buttons,
-                                 chat_id)  # вызываем три кнопки с одним описанием
-                text = None
-            if text == '/lang ru':
-                lang_all.update({chat_id: 'ru'})
-                bot.send_message('Текст будет переводиться на Русский', chat_id)
-                text = None
-            if text == '/lang en':
-                lang_all.update({chat_id: 'en'})
-                bot.send_message('Text will be translated into English', chat_id)
-                text = None
-            if text == '/lang auto':
-                lang_all.update({chat_id: 'auto'})
-                bot.send_message('Русский|English - автоматически|automatically', chat_id)
-                text = None
-            if payload is not None:
-                lang_all.update({chat_id: payload})
-                lang = lang_all.get(chat_id)
-                text = None
-                if lang == 'ru':
-                    bot.send_answer_callback(callback_id, 'Текст будет переводиться на Русский')
-                    bot.delete_message(mid)
-                elif lang == 'auto':
-                    bot.send_answer_callback(callback_id, 'Русский|English - автоматически|automatically')
-                    bot.delete_message(mid)
-                else:
-                    bot.send_answer_callback(callback_id, 'Text will be translated into English')
-                    bot.delete_message(mid)
+            name = bot.get_name(last_update)
+            admins = bot.get_chat_admins(chat_id)
+            if not admins or admins and name in [i['name'] for i in admins['members']]:
+                if text == '/lang' or text == '@gotranslatebot /lang':
+                    buttons = [[{"type": 'callback',
+                                 "text": 'Авто|Auto',
+                                 "payload": 'auto'},
+                                {"type": 'callback',
+                                 "text": 'Русский',
+                                 "payload": 'ru'},
+                                {"type": 'callback',
+                                 "text": 'English',
+                                 "payload": 'en'}]]
+                    bot.send_buttons('Направление перевода\nTranslation direction', buttons,
+                                     chat_id)  # вызываем три кнопки с одним описанием
+                    text = None
+                if text == '/lang ru' or text == '@gotranslatebot /lang ru':
+                    lang_all.update({chat_id: 'ru'})
+                    bot.send_message('Текст будет переводиться на Русский', chat_id)
+                    text = None
+                if text == '/lang en' or text == '@gotranslatebot /lang en':
+                    lang_all.update({chat_id: 'en'})
+                    bot.send_message('Text will be translated into English', chat_id)
+                    text = None
+                if text == '/lang auto' or text == '@gotranslatebot /lang auto':
+                    lang_all.update({chat_id: 'auto'})
+                    bot.send_message('Русский|English - автоматически|automatically', chat_id)
+                    text = None
+                if payload is not None:
+                    lang_all.update({chat_id: payload})
+                    lang = lang_all.get(chat_id)
+                    text = None
+                    if lang == 'ru':
+                        bot.send_answer_callback(callback_id, 'Текст будет переводиться на Русский')
+                        bot.delete_message(mid)
+                    elif lang == 'auto':
+                        bot.send_answer_callback(callback_id, 'Русский|English - автоматически|automatically')
+                        bot.delete_message(mid)
+                    else:
+                        bot.send_answer_callback(callback_id, 'Text will be translated into English')
+                        bot.delete_message(mid)
+
             if type_upd == 'bot_started':
                 bot.send_message(
                     'Отправьте или перешлите боту текст. Язык переводимого текста определяется автоматически. '
@@ -118,15 +122,15 @@ def main():
                 text_const = bot.get_construct_text(last_update)
                 sid = bot.get_session_id(last_update)
                 if text_const:
-                    translt, lang_detect = translate(text_const, 'auto')
+                    translt, lang_detect = translate(text_const, 'en')
                     if translt:
                         bot.send_construct_message(sid, hint=None, text=translt)
                     else:
-                        bot.send_construct_message(sid, 'Введите текст для перевода и отправки в чат | '
-                                                        'Enter the text to be translated and send to the chat')
+                        bot.send_construct_message(sid, 'Введите текст для перевода на Английский и отправки в чат | '
+                                                        'Enter the text to be translated into English and sent to chat')
                 else:
-                    bot.send_construct_message(sid, 'Введите текст для перевода и отправки в чат | '
-                                                    'Enter the text to be translated and send to the chat')
+                    bot.send_construct_message(sid, 'Введите текст для перевода на Английский и отправки в чат | '
+                                                    'Enter the text to be translated into English and send to chat')
             elif text:
                 translt, lang_detect = translate(text, lang)
                 if translt:
@@ -142,7 +146,6 @@ def main():
                             bot.send_message(translt, chat_id)
                     else:
                         bot.send_message('Перевод невозможен\nTranslation not available', chat_id)
-        continue
 
 
 if __name__ == '__main__':
